+++ .travis.yml
@@ -9,7 +9,7 @@ deploy:
   provider: rubygems
   api_key:
     secure: 1AilejnVgIXffKNfZIkGIlRNTFMK3ad3Fq4RwR/nE5ORFWoso36ojobAAEHfImBM6lQX98KSKNRMH/37S+DzzVSKqgvBIGAyto7SsCOOYM5HEj5z5xnEe11mxz3z/+Sfi2aNEDVkrHz18S68oZnm37UghF13f2UuNuTUhsiiF8PWbWgcu1flewmGiuuiGvQHsoioLJGXO538xnj4Vr+V/iTFQgnn4VaoOwpOrBdoY7gg81UYDYO0+lxHJ9DCY/7QYQseDH8BCXDrndD75hrMFhzcuy9p9PLZ1d4uudS56B7DtH7PgmL9qPzl/SL5kZk4jlJw1i9Y8RP/sI8nxmLpFgJ3kmGSsWm3YGEhnwpyMd48hyybzgltKuzsqB1vnAkvTtPWX5SwWjUp2pbaqrduEiMdLQ8S35ViBToNpmtul5B36vQZ+EpMhZukQ1dN8jzKffJh7NCtsW0xpXsI70KZLPwP+KO6IvE9dv/AA16XF4tcy7/AF7pxyh49ao9xSX1IzIKWAxOA/WOaOrEYoPvLg/I9N9XAUK/W/MUKTDdLmNQsiWqcw1Ub7gPWlZw5oCJY6aV8CXDushWHWkqTmMs7pBgfpQ5JIhNrvzrYNyf+7QAnCdJ6nCr99BAyTTy5RgiI4P/hGwPEUpYVJe2qB4AQZbZiLpBzAfTGDipmUypzIg0=
-  gem: mumukit-core
+  gem: patch_adams
   on:
     tags: true
-    repo: mumuki/mumukit-core
+    repo: mumuki/patch_adams
