+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in mumukit-core.gemspec
+# Specify your gem's dependencies in patch_adams.gemspec
 gemspec
