--- lib/mumukit/core.rb
@@ -1,25 +1,0 @@
-require 'mumukit/core/version'
-require 'active_support/all'
-
-module Mumukit
-  module Core
-  end
-end
-
-require_relative './core/status'
-require_relative './core/object'
-require_relative './core/module'
-require_relative './core/json'
-require_relative './core/hash'
-require_relative './core/string'
-require_relative './core/nil'
-require_relative './core/array'
-require_relative './core/regexp'
-require_relative './core/request'
-require_relative './core/configurable'
-require_relative './core/modulability'
-require_relative './core/variability'
-require_relative './core/likeability'
-require_relative './core/file'
-require_relative './core/i18n'
-require_relative './core/yaml'
