--- lib/mumukit/core/version.rb
@@ -1,5 +1,0 @@
-module Mumukit
-  module Core
-    VERSION = '1.14.0'
-  end
-end
