+++ lib/patch_adams.rb
@@ -1,0 +1,23 @@
+require 'patch_adams/version'
+require 'active_support/all'
+
+module PatchAdams
+end
+
+require_relative './patch_adams/status'
+require_relative './patch_adams/object'
+require_relative './patch_adams/module'
+require_relative './patch_adams/json'
+require_relative './patch_adams/hash'
+require_relative './patch_adams/string'
+require_relative './patch_adams/nil'
+require_relative './patch_adams/array'
+require_relative './patch_adams/regexp'
+require_relative './patch_adams/request'
+require_relative './patch_adams/configurable'
+require_relative './patch_adams/modulability'
+require_relative './patch_adams/variability'
+require_relative './patch_adams/likeability'
+require_relative './patch_adams/file'
+require_relative './patch_adams/i18n'
+require_relative './patch_adams/yaml'
