+++ lib/patch_adams/activemodel.rb
