+++ lib/patch_adams/array.rb
