+++ lib/patch_adams/file.rb
