+++ lib/patch_adams/hash.rb
