+++ lib/patch_adams/json.rb
