+++ lib/patch_adams/likeability.rb
