+++ lib/patch_adams/modulability.rb
