+++ lib/patch_adams/module.rb
