+++ lib/patch_adams/nil.rb
