+++ lib/patch_adams/object.rb
