+++ lib/patch_adams/regexp.rb
