+++ lib/patch_adams/rspec.rb
