+++ lib/patch_adams/status.rb
