+++ lib/patch_adams/string.rb
