+++ lib/patch_adams/variability.rb
