+++ lib/patch_adams/version.rb
@@ -1,0 +1,3 @@
+module PatchAdams
+  VERSION = '1.14.0'
+end
