+++ lib/patch_adams/yaml.rb
