+++ spec/patch_adams/activemodel_spec.rb
