+++ spec/patch_adams/arrays_spec.rb
