+++ spec/patch_adams/hash_spec.rb
