+++ spec/patch_adams/json_spec.rb
