+++ spec/patch_adams/module_spec.rb
