+++ spec/patch_adams/request_spec.rb
