+++ spec/patch_adams/rspec_spec.rb
