+++ spec/patch_adams/status_spec.rb
