+++ spec/patch_adams/yaml_spec.rb
