+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 require 'codeclimate-test-reporter'
 SimpleCov.start
 
-require 'mumukit/core/rspec'
-require 'mumukit/core'
-require 'mumukit/core/request'
-require 'mumukit/core/activemodel'
+require 'patch_adams/rspec'
+require 'patch_adams'
+require 'patch_adams/request'
+require 'patch_adams/activemodel'
